moved registry.go to discovery
