+++ internal/controller/discovery/registry.go
@@ -1,4 +1,4 @@
-package registry
+package discovery
 
 import (
 	"fmt"
